last correction before submit

+++ ft_strrchr.c
@@ -6,14 +6,14 @@
 /*   By: nhan <necat.han42@gmail.com>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/23 14:17:32 by nhan              #+#    #+#             */
-/*   Updated: 2023/12/08 16:56:01 by nhan             ###   ########.fr       */
+/*   Updated: 2023/12/08 17:16:47 by nhan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
 
 char	*ft_strrchr(const char *s, int c)
 {
-	size_t	i;
+	int	i;
 
 	i = ft_strlen(s);
 	while (i >= 0)
